Rearrange file structure
