--- src/components/Skillspage/Frameworks/libraries.js
@@ -1,29 +1,0 @@
-import rails from "./../../../assets/rails.svg"
-import react from "./../../../assets/react.svg"
-import redux from "../../../assets/redux.svg"
-import postcss from "../../../assets/postcss.svg"
-
-const libraries = [
-  {
-    id: 0,
-    icon: react,
-    name: 'React'
-  },
-  {
-    id: 1,
-    icon: redux,
-    name: 'Redux'
-  },
-  {
-    id: 2,
-    icon: postcss,
-    name: 'PostCSS'
-  },
-  {
-    id: 3,
-    icon: rails,
-    name: 'Ruby on Rails'
-  },
-]
-
-export default libraries;
