--- src/components/Skillspage/Languages/languagesArr.js
@@ -1,35 +1,0 @@
-import javascript from '../../../assets/javascript.svg'
-import html from '../../../assets/html.svg'
-import css from '../../../assets/css.svg'
-import ruby from '../../../assets/ruby.svg'
-import postgres from '../../../assets/postgres.svg'
-
-const languageArr = [
-  {
-    id: 0,
-    icon: html,
-    name: 'HTML',
-  },
-  {
-    id: 1,
-    icon: css,
-    name: 'CSS',
-  },
-  {
-    id: 2,
-    icon: javascript,
-    name: 'Javascript',
-  },
-  {
-    id: 3,
-    icon: ruby,
-    name: 'Ruby',
-  },
-  {
-    id: 4,
-    icon: postgres,
-    name: 'PostgreSQL',
-  },
-]
-
-export default languageArr;
