+++ src/components/Skillspage/Tools.jsx
@@ -1,9 +1,9 @@
-import toolKit from "./toolkit";
-import "../icons.css"
+import toolKit from "../../data/toolkit";
+import "./icons.css"
 
 function Tools() {
   return (
-    <ul className="flex animate box_shadow icons-backdrop">
+    <ul className="flex box_shadow icons-backdrop">
       <h3>Tools</h3>
       <div className="flex icons-container">
         {toolKit.map((tool) => (
