--- src/components/Skillspage/Tools/toolkit.js
@@ -1,41 +1,0 @@
-import github from  "./../../../assets/github1.svg"
-import webpack from  "./../../../assets/webpack.svg"
-import vscode from  "./../../../assets/vscode.svg"
-import jest from  "./../../../assets/jest.svg"
-import rspec from  "./../../../assets/rspec.svg"
-import git from  "./../../../assets/git.svg"
-
-const toolKit = [
-  {
-    id: 0,
-    icon: git,
-    name: 'Git'
-  },
-  {
-    id: 1,
-    icon: github,
-    name: 'Github'
-  },
-  {
-    id: 2,
-    icon: webpack,
-    name: 'Webpack'
-  },
-  {
-    id: 3,
-    icon: vscode,
-    name: 'VScode'
-  },
-  {
-    id: 4,
-    icon: jest,
-    name: 'Jest'
-  },
-  {
-    id: 5,
-    icon: rspec,
-    name: 'Rspec'
-  }
-]
-
-export default toolKit;
