+++ src/components/Skillspage/icons.css
@@ -5,7 +5,6 @@
   padding: 1rem;
   gap: 1.3rem;
   width: 100%;
-  max-width: 15rem;
   background-color: var(--secondary-color);
 }
 
@@ -21,6 +20,7 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  max-width: 3rem;
   gap: 0.5rem;
   position: relative;
 }
@@ -28,7 +28,7 @@
 .backdrop {
   align-items: center;
   max-width: 3rem;
-  padding: 0.4rem;
+  padding: 0.7rem;
   border-radius: 35%;
   background-color: var(--primary-color);
 }
@@ -37,14 +37,31 @@
   width: 100%;
 }
 
-@media screen and (min-width: 1180px) {
+@media screen and (min-width: 580px) {
+  .icons-backdrop {
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .icons-container {
+    width: unset;
+  }
+}
+
+@media screen and (min-width: 800px) {
   .backdrop {
     max-width: 3rem;
     padding: 0.7rem;
   }
 
+  .icon-container {
+    max-width: 5rem;
+  }
+
   .icons-backdrop {
+    justify-content: unset;
+    flex-direction: column;
     max-width: 20rem;
-    height: 15.6rem;
+    height: 16.6rem;
   }
 }
