+++ src/components/Skillspage/skillspage.module.css
@@ -1,11 +1,13 @@
-.skillspage {
-  flex-wrap: nowrap;
-}
-
 .components {
   justify-content: center;
   align-items: center;
+  flex-direction: column;
   width: 100%;
-  flex-wrap: wrap;
   gap: 1rem;
+}
+
+@media screen and (min-width: 800px) {
+  .components {
+    flex-direction: row;
+  }
 }
